[TLE][MTHREADS] Add backend-local TLE barrier support
